refactor(qr): move payload classification out of presentation into payload-v2
presentation.ts formats dates, fingerprints, and frame positions; classifying
protocol prefixes is the QR layer's job, and payload-v2.ts already owns that
via classifyV2Payload. The predicate is not classifyV2Payload itself: the
reserved OCB2 prefix classifies but must never display as valid.
