+++ src/components/qr-display.tsx
@@ -2,7 +2,7 @@ import { useEffect, useRef, useState, type ReactNode } from "react"
 import { AlertCircle, Expand, X } from "lucide-react"
 import { renderQrDataUrl } from "@/qr/encode"
 import type { QrEcLevel } from "@/schemas/domain"
-import { isQrCryptPayload } from "@/features/presentation"
+import { isQrCryptPayload } from "@/qr/payload-v2"
 import { toAppError } from "@/crypto/errors"
 import { cn } from "@/lib/utils"
 import { Alert, AlertDescription, AlertTitle } from "@/components/ui/alert"
