+++ src/features/presentation.ts
@@ -1,11 +1,5 @@
 import { translate, type Language } from "@/i18n/messages"
-import { QR_PREFIX_V2 } from "@/qr/payload-v2"
 import type { UiAlgorithm } from "@/schemas/domain"
-
-// OCB2 is reserved and rejected by every decoder; the display side must not call it valid.
-const QR_CRYPT_PREFIXES = Object.entries(QR_PREFIX_V2)
-  .filter(([kind]) => kind !== "encrypted-seed-backup")
-  .map(([, prefix]) => prefix)
 
 export const ALGORITHM_LABELS: Record<
   Language,
@@ -81,8 +75,4 @@ export function formatFramePositions(
           }),
         )
         .join(translate(language, "presentation.frameSeparator"))
-}
-
-export function isQrCryptPayload(payload: string): boolean {
-  return QR_CRYPT_PREFIXES.some((prefix) => payload.startsWith(prefix))
 }
