+++ tests/unit/presentation.test.ts
@@ -1,10 +1,5 @@
 import { describe, expect, it } from "vitest"
-import {
-  formatDateTime,
-  formatFingerprint,
-  isQrCryptPayload,
-} from "@/features/presentation"
-import { QR_PREFIX_V2 } from "@/qr/payload-v2"
+import { formatDateTime, formatFingerprint } from "@/features/presentation"
 
 describe("presentation formatting", () => {
   it("returns the em-dash fallback for a non-finite timestamp", () => {
@@ -18,15 +13,5 @@ describe("presentation formatting", () => {
 
   it("returns short input unchanged instead of fabricating groups", () => {
     expect(formatFingerprint("abc")).toBe("abc")
-  })
-
-  it("recognizes only active v2 prefixes", () => {
-    const { "encrypted-seed-backup": reserved, ...active } = QR_PREFIX_V2
-    for (const prefix of Object.values(active)) {
-      expect(isQrCryptPayload(`${prefix}payload`)).toBe(true)
-    }
-    expect(isQrCryptPayload(`${reserved}payload`)).toBe(false)
-    expect(isQrCryptPayload("OCX9:payload")).toBe(false)
-    expect(isQrCryptPayload("plain text")).toBe(false)
   })
 })
